Rename transform files
