+++ src/style/pauseable_placement.ts
@@ -1,6 +1,6 @@
 import {browser} from '../util/browser';
 import {Placement} from '../symbol/placement';
-import type {ITransform} from '../geo/transform';
+import type {ITransform} from '../geo/transform_interface';
 import type {StyleLayer} from './style_layer';
 import type {SymbolStyleLayer} from './style_layer/symbol_style_layer';
 import type {Tile} from '../source/tile';
