+++ src/style/style_layer/line_style_layer.ts
@@ -13,7 +13,7 @@ import {isZoomExpression, Step} from '@maplibre/maplibre-gl-style-spec';
 import type {FeatureState, LayerSpecification} from '@maplibre/maplibre-gl-style-spec';
 import type {Bucket, BucketParameters} from '../../data/bucket';
 import type {LineLayoutProps, LinePaintProps} from './line_style_layer_properties.g';
-import type {ITransform} from '../../geo/transform';
+import type {ITransform} from '../../geo/transform_interface';
 import type {VectorTileFeature} from '@mapbox/vector-tile';
 
 export class LineFloorwidthProperty extends DataDrivenProperty<number> {
