+++ src/ui/handler/transform-provider.ts
@@ -1,6 +1,6 @@
 import type {Map} from '../map';
 import type {PointLike} from '../camera';
-import type {ITransform} from '../../geo/transform';
+import type {ITransform} from '../../geo/transform_interface';
 import Point from '@mapbox/point-geometry';
 import {LngLat} from '../../geo/lng_lat';
