+++ src/util/smart_wrap.ts
@@ -1,7 +1,7 @@
 import {LngLat} from '../geo/lng_lat';
 
 import type Point from '@mapbox/point-geometry';
-import type {ITransform} from '../geo/transform';
+import type {ITransform} from '../geo/transform_interface';
 
 /**
  * Given a LngLat, prior projected position, and a transform, return a new LngLat shifted
